requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.1.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
coverage==7.6.12
pytest-cov==6.0.0
